+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmcrosslink.map.inc.php
@@ -91,7 +91,7 @@ $xpdo_meta_map['rmCrosslink']= array (
     'Author' => 
     array (
       'class' => 'modUser',
-      'local' => 'author_id',
+      'local' => 'createdby',
       'foreign' => 'id',
       'cardinality' => 'one',
       'owner' => 'foreign',