+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmcrosslinkrelated.map.inc.php
@@ -163,7 +163,7 @@ $xpdo_meta_map['rmCrosslinkRelated']= array (
     'Author' => 
     array (
       'class' => 'modUser',
-      'local' => 'author_id',
+      'local' => 'createdby',
       'foreign' => 'id',
       'cardinality' => 'one',
       'owner' => 'foreign',