+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmnoteissue.map.inc.php
@@ -159,7 +159,7 @@ $xpdo_meta_map['rmNoteIssue']= array (
     'Author' => 
     array (
       'class' => 'modUser',
-      'local' => 'author_id',
+      'local' => 'createdby',
       'foreign' => 'id',
       'cardinality' => 'one',
       'owner' => 'foreign',