+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmoption.map.inc.php
@@ -103,15 +103,15 @@ $xpdo_meta_map['rmOption']= array (
         ),
       ),
     ),
-    'position' => 
+    'pos' => 
     array (
-      'alias' => 'position',
+      'alias' => 'pos',
       'primary' => false,
       'unique' => false,
       'type' => 'BTREE',
       'columns' => 
       array (
-        'position' => 
+        'pos' => 
         array (
           'length' => '',
           'collation' => 'A',