+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmtimeline.map.inc.php
@@ -141,7 +141,7 @@ $xpdo_meta_map['rmTimeline']= array (
     'Author' => 
     array (
       'class' => 'modUser',
-      'local' => 'author_id',
+      'local' => 'createdby',
       'foreign' => 'id',
       'cardinality' => 'one',
       'owner' => 'foreign',