+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmtimelineproject.map.inc.php
@@ -176,7 +176,7 @@ $xpdo_meta_map['rmTimelineProject']= array (
     'Author' => 
     array (
       'class' => 'modUser',
-      'local' => 'author_id',
+      'local' => 'createdby',
       'foreign' => 'id',
       'cardinality' => 'one',
       'owner' => 'foreign',